// frontend/client/types.gen.ts
// @ts-nocheck
// @ts-nocheck
// This file is auto-generated by @hey-api/openapi-ts

export type EvalLog = {
  explanation?: string | null
}

export type EvalResult = {
  score?: number | null
  triggered?: boolean | null
  triggered_escalation?: boolean | null
  triggered_guardrail?: boolean | null
  log?: EvalLog | null
}

export type HTTPValidationError = {
  detail?: Array<ValidationError>
}

export type Message_Input = {
  role: MessageRole
  content: string
  metadata?: MessageMetadata_Input
}

export type Message_Output = {
  role: MessageRole
  content: string
  metadata?: MessageMetadata_Output
}

export type MessageMetadata_Input = {
  original_llm_response?: string | null
  is_expert_answer?: boolean | null
  guardrailed?: boolean | null
  escalated_to_sme?: boolean | null
  scores?: {
    [key: string]: EvalResult
  } | null
}

export type MessageMetadata_Output = {
  original_llm_response?: string | null
  is_expert_answer?: boolean | null
  guardrailed?: boolean | null
  escalated_to_sme?: boolean | null
  scores?: {
    [key: string]: EvalResult
  } | null
}

export type MessageRole = 'user' | 'assistant'

export type ValidationError = {
  loc: Array<string | number>
  msg: string
  type: string
}

export type ChatApiChatPostData = {
  body: Message_Input
}

export type ChatApiChatPostResponse = Message_Output

export type ChatApiChatPostError = HTTPValidationError